backend/pyproject.toml
[tool.poetry]
name = "di-cho-tien-loi-backend"
version = "0.1.0"
description = "FastAPI backend for the Di Cho Tien Loi multi-platform application"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
pydantic-settings = "^2.2.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
email-validator = "^2.1.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.3"
celery = "^5.3.6"
boto3 = "^1.34.0"
minio = "^7.2.7"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-asyncio = "^0.23.0"
factory-boy = "^3.3.0"
ruff = "^0.3.0"
black = "^24.2.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
app = "app.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
